1, fix update business broker net

+++ database/migrations/2019_07_02_075921_business_broker_net.php
@@ -42,7 +42,9 @@ class BusinessBrokerNet extends Migration
      */
     public function down()
     {
-        Schema::drop('business_broker_net');
-        Schema::drop('business_broker_net_member');
+        if(env('APP_ENV') == 'local'){
+            Schema::drop('business_broker_net');
+            Schema::drop('business_broker_net_member');
+        }
     }
 }
